+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <com.bcc.opensource.squaremenu.SquareMenu
+    <com.filkom.bcc.gridmenu.GridMenu
         android:id="@+id/square_menu"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
