+++ gridmenulibrary/src/main/res/values/attrs.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="GridMenu">
+        <attr name="spanCount" format="integer" />
+    </declare-styleable>
+</resources>
