--- squaremenu/src/main/res/values/attrs.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <declare-styleable name="SquareMenu">
-        <attr name="spanCount" format="integer" />
-    </declare-styleable>
-</resources>
